backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
pytest
httpx
pydantic[email]
slowapi
python-multipart
google-cloud-logging
google-cloud-secret-manager
feedparser
google-generativeai
email-validator
